% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_library_version.R
\name{apply_library_version}
\alias{apply_library_version}
\title{Adjust package-loading in R scripts to pama-style}
\usage{
apply_library_version(
  req.file.path = getwd(),
  req.file.name = "requirements.txt",
  lists = TRUE,
  script.paths = NULL,
  patterns = c("library\\\\(([^,)]+)\\\\)", "require\\\\(([^,)]+)\\\\)")
)
}
\arguments{
\item{req.file.path}{(chr vector): Path in which to create the requirements file}

\item{req.file.name}{(chr vector): Name of the requirements-file}

\item{lists}{(bool): Insert a list structure into the file? TRUE=yes, FALSE=no}

\item{script.paths}{(chr vector): Provide a path / paths to the R scripts that should be adjusted}

\item{patterns}{(chr vector): \href{https://cran.r-project.org/web/packages/stringr/vignettes/regular-expressions.html}{RegEx} to detect package-loadings in the r-scripts}
}
\description{
Adjust package-loading in R scripts to pama-style
}
\details{
Adjusting the package-loading in R scripts to pama-style
}
\section{Dependencies}{

- get_requirements()
}

\section{Side effects}{
 Adjusting the package-loading in R scripts to pama-style
}

\section{Return}{
 None
}

\examples{
\dontrun{
#Creating a req-file in the working directory called "draft-requirements.txt", with list structure
apply_library_version()

#Adjust package-loading within R scripts in .\Scripts to pama-style, using "requirements.txt" within .\
apply_library_version(req.file.path = r"(C:\Users\sress\Desktop\titanic-r-master)",
                      req.file.name = "requirements.txt",
                      script.paths = r"(C:\Users\sress\Desktop\titanic-r-master\Scripts)")
}

}
\author{
Simon Ress
}
\keyword{library_version}
\keyword{pama-style,}
