% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_package_url.R
\name{build_package_url}
\alias{build_package_url}
\title{Build the URL of package-versions}
\usage{
build_package_url(
  .main = TRUE,
  .repo = parent.frame()$repo,
  .overview = F,
  .package = parent.frame()$package,
  .version = parent.frame()$version,
  .cran.mirror = parent.frame()$cran.mirror,
  .main.path = parent.frame()$main.path,
  .archiv.path = parent.frame()$archiv.path
)
}
\arguments{
\item{.main}{(bool):
TRUE (default) = Build the URL of an current package version in the repo, `.main.path` is used.
FALSE = TRUE=Build the URL of an older package version in the repos archive, `.archive.path` is used.}

\item{.repo}{(chr value): Which type of repo is used? This information is utilized when building the URL,
because each repo has its own file structure. Currently supported: "cran" & "nexus"}

\item{.overview}{(bool):
FALSE (default)= Return link to package-version src (.tar.gz ) or bin (.zip) file
TRUE=Return page with overview of all versions of a package (e.g.: https://cloud.r-project.org/src/contrib/Archive/dplyr/)}

\item{.package}{(chr value): Name of the package to install (e.g. "ggplot2")}

\item{.version}{(chr value): Version of the package to install (e.g. "3.4.0")}

\item{.cran.mirror}{(chr value): Main url of the repo to use (e.g. "https://cloud.r-project.org/")}

\item{.main.path}{(chr value): URL-path to the pages main page of the repo to use (e.g. "src/contrib/"")}

\item{.archiv.path}{(chr value): URL-path to the archive of the repo to use (e.g. "src/contrib/Archive/")}
}
\description{
Creates the URL of a package file for installation.
This URL build on the given URL-components (`.cran.mirror`, `.main.path` & `.archiv.path`),
the packages information (`.package` & `.version`), the packages storage location in archive or main (`.main`) and
the type of given CRAN repo (`.repo`).
}
\note{
Here the structure of cran:
Packages *main* page: https://cran.r-project.org/web/packages/dplyr/index.html
Packages *main* storage location: https://cran.r-project.org/src/contrib/dplyr_1.1.2.tar.gz <-
Packages *archive* page: https://cran.r-project.org/src/contrib/Archive/dplyr/
Packages *archive* storage location: https://cran.r-project.org/src/contrib/Archive/dplyr/dplyr_1.1.0.tar.gz
}
\section{Dependencies}{
 None
}

\section{Side effects}{
 None
}

\section{Return}{
 (chr value) URL of the package version
}

\examples{
\dontrun{
# Settings:
.package = "dbplyr"
.version = "2.3.1"
.cran.mirror = "https://cran.r-project.org/"
.main.path = "src/contrib/"
archiv.path = "src/contrib/Archive/"

# Build an URL of a current package version on CRAN
build_package_url(.repo = "cran")
   #Output should be: "https://cran.r-project.org/src/contrib/dbplyr_2.3.1.tar.gz"

# Build an URL of the CRAN archive
build_package_url(.main=FALSE, .repo = "cran")
   # Output should be: "https://cran.r-project.org/src/contrib/Archive/dbplyr/dbplyr_2.3.1.tar.gz"

# Build an URL of the "nexus" repo type
.repo = "nexus"
.cran.mirror = "https://<***>.de/service/rest/repository/browse/r-public/"
.archiv.path = "bin/windows/contrib/"
build_package_url(.repo = repo)
   # Output should be: "https://<***>.de/service/rest/repository/browse/r-public/bin/windows/contrib/4.2/dbplyr/2.3.1/dbplyr_2.3.1.zip"

# Test installation of a package with the created URL
.cran.mirror = "https://cran.r-project.org/"
.main.path = "src/contrib/"
.url = build_package_url(.main=T, .repo = "cran", .package="gt", .version="0.9.0")
install.packages(url, repos=NULL, type="source")
library(gt)
remove.packages("gt")

}

}
\seealso{
`browseVignettes("pama")`, `help(package = "pama")`
}
\author{
Simon Ress
}
\keyword{URL}
\keyword{cran}
\keyword{package-versions}
