% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations2.R
\name{citations2}
\alias{citations2}
\title{Creation of a 'citation table' for packages}
\usage{
citations2(
  use.requirements.file = TRUE,
  requirements.file.name = "requirements.txt",
  requirements.file.path = getwd(),
  packages = c("ggplot2", "dplyr"),
  output.path = getwd()
)
}
\arguments{
\item{use.requirements.file}{Whether (TRUE) to use a requirements-file as source of packages to be cited or (FALSE) to provide package name in 'packages'-attribute}

\item{requirements.file.name}{Name of the requirements-file}

\item{requirements.file.path}{Path of the requirements-file}

\item{packages}{If requirements-file is not used: Vector with the names of the packages to be cited}

\item{output.path}{Where to save the csv-file containing the citation table}
}
\description{
Creation of a 'citation table' for packages
}
\details{
Creating a citation table the given packages including the columns 'Package', 'Version' and 'Zitat'.
}
\section{Dependencies}{
 get_requirements()
}

\section{Side effects}{
 Saving the citation table for packages as .csv-file within the output.path. File name: paste0("/lib_citations_",Sys.Date(),".csv")
}

\section{Return}{
 Data frame including the citation table for packages
}

\examples{
\dontrun{
# Citing the packages from the "requirements.txt" within the current working directory
citations2()

# Citing the packages from the "req.txt" within "C:/GIT/Packages/PaMaTo" and saving ourput within the current working directory
citations2(requirements.file.path = "C:/GIT/Packages/PaMaTo",
           requirements.file.name = "req.txt")
}

}
\seealso{
\code{\link[pama]{get_requirements}}
}
\author{
Simon Ress
}
\keyword{citation}
\keyword{packages}
