% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_package_version_on_cran.r
\name{find_package_version_on_cran}
\alias{find_package_version_on_cran}
\title{Find the URL to a specific package version on CRAN}
\usage{
find_package_version_on_cran(
  package,
  version,
  cran.mirror = "https://cloud.r-project.org/",
  archiv.path = "src/contrib/Archive/",
  main.path = "src/contrib/",
  repo = "cran"
)
}
\arguments{
\item{package}{(chr value): Name of the package to install (e.g. "ggplot2")}

\item{version}{(chr value): Version of the package to install (e.g. "3.4.0")}

\item{cran.mirror}{(chr value): Main url of the cran mirror to use (e.g. "https://cloud.r-project.org/")}

\item{archiv.path}{(chr value): URL-path to the archive of the cran mirror to use (e.g. "src/contrib/Archive/")}

\item{main.path}{(chr value): URL-path to the pages main page of the cran mirror to use (e.g. "src/contrib/"")}

\item{repo}{(chr value): Which type of repo is used? This information is utilized when building the URL,
because each repo has its own file structure. Currently supported: "cran" & "nexus"}
}
\value{
(chr vector):
return[1] (chr): URL of the package version or an error message

return[2] (chr): Version of the package (sometimes the version changes, therefore an output is needed)
}
\description{
Finding the URL to a specific package version on CRAN. Therefore searching the archive and the main page of the given cran instance.
Then trying to changing the version name structure by replacing the last "." by "-" and deleting trailing zeros.

(e.g. "0.1.10" to "0.1-1", because \href{https://cloud.r-project.org/src/contrib/Archive/dplyr/}{dplyr_0.8.0} requires "plogr_0.1.10"
but in \href{https://cloud.r-project.org/src/contrib/Archive/plogr/}{plogrs archive} this version is named "plogr_0.1-1]
}
\note{
See all available CRAN Packages by Name here: https://cran.r-project.org/web/packages/available_packages_by_name.html
}
\section{Dependencies}{
 None
}

\section{Side effects}{
 None
}

\examples{
\dontrun{
.out = find_package_version_on_cran(package = package, version = version)
package.url = .out[1]
version = .out[2]
}

}
\keyword{CRAN}
\keyword{finding}
\keyword{on}
\keyword{package-versions}
