% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Get the dependencies of an package on CRAN}
\usage{
get_dependencies(
  package,
  version,
  cran.mirror = "https://cloud.r-project.org/",
  archiv.path = "src/contrib/Archive/",
  main.path = "src/contrib/"
)
}
\arguments{
\item{package}{sad}

\item{version}{sd}

\item{cran.mirror}{vv}

\item{archiv.path}{fdf}

\item{main.path}{xc}
}
\description{
Extracts all dependencies (package names and versions) of the giving version of the package
}
\note{
See all available CRAN Packages by Name here: https://cran.r-project.org/web/packages/available_packages_by_name.html
}
\section{Dependencies}{
 None
}

\section{Side effects}{
 None
}

\section{Return}{
 List of lists with structure c(<package_name>, <version>)
}

\examples{
\dontrun{

#package-VERSION does not exist -> returns "version-error"
get_dependencies("ggmap","2.6.0" )

# package-VERSION does not exist -> returns "version-error"
get_dependencies("ggplot2","26.6.6" )

# package-NAME does not exist -> returns "package-name-error"
get_dependencies("ggmapAAA","3.0.0" )

# package-NAME does not exist -> returns "package-name-error"
get_dependencies("ggplot222","1.0.0" )

# package version in archive (-> https://cran.r-project.org/src/contrib/Archive/ggplot2/)
get_dependencies("ggplot2","3.1.1" )

# current version, date: 2023.03.08 (on main page -> https://cran.r-project.org/web/packages/ggplot2/index.html)
get_dependencies("ggplot2","3.4.1" )

# package version in archive (->https://cran.r-project.org/src/contrib/Archive/ggmap/)
get_dependencies("ggmap","3.0.0" )

# current version, date: 2023.03.08 (on main page -> https://cloud.r-project.org/web/packages/ggmap/)
get_dependencies("ggmap","3.0.1" )
}

}
\author{
Simon Ress
}
