% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_installed_packages.R
\name{get_installed_packages}
\alias{get_installed_packages}
\title{Returns installed packages and their versions}
\usage{
get_installed_packages(lib.search.path = .libPaths())
}
\arguments{
\item{lib.search.path}{(chr vector): Paths to search for packages, by default .libPaths() <- Internal search paths of R}
}
\description{
Returns installed packages and their versions
}
\section{Dependencies}{
 - update_packages_search_path()
}

\section{Side effects}{
 None
}

\section{Return}{
 List of lists with structure c(<package_name>, <version>)
}

\examples{
get_installed_packages()

get_installed_packages("C:/Program Files/R/R-4.2.2/library")

}
\seealso{
\code{\link[utils]{.libPaths()}}
}
\author{
Simon Ress
}
\keyword{Returns}
\keyword{and}
\keyword{installed}
\keyword{packages}
\keyword{versions}
