% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_requirements.R
\name{get_requirements}
\alias{get_requirements}
\title{Get data from requirements-file}
\usage{
get_requirements(req.file.name = "requirements.txt", req.file.path = getwd())
}
\arguments{
\item{req.file.name}{(chr vector): Name of the requirement-file}

\item{req.file.path}{(chr vector): Folder of the requirement-file}
}
\description{
Get data from requirements-file
}
\details{
Returns the information of the requirements-file
}
\section{Dependencies}{
 NONE
}

\section{Side effects}{
 None
}

\section{Return}{
 List of lists. First list ('r.version') returns the required r-version. Other list (e.g. 'main') return vectors including required versions of packages in format: "<package_name> <version>"
}

\examples{
\dontrun{
# Load requirements specifies in "requirements.txt" within the working directory
get_requirements()
}

}
\seealso{
\code{\link[base]{.libPaths}}
}
\author{
Simon Ress
}
\keyword{requirements-file}
