% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_requirements.R
\name{load_requirements}
\alias{load_requirements}
\title{...}
\usage{
load_requirements(
  req.file.path = getwd(),
  req.file.name = "requirements.txt",
  list = "all",
  library.folder.path = getwd(),
  library.folder.name = "lib"
)
}
\arguments{
\item{req.file.path}{(chr vector):}

\item{req.file.name}{(chr vector):}

\item{list}{(chr vector):}

\item{library.folder.path}{(chr vector):}

\item{library.folder.name}{(chr vector):}
}
\description{
...
}
\details{
test
}
\section{Dependencies}{

- update_packages_search_path()

- get_requirements()

- library_version()
}

\section{Side effects}{
 ...
}

\section{Return}{
 ...
}

\examples{
\dontrun{
# load all packages specified in the 'requirements.txt.' within the current working directory
detach_none_base() # detach all none base packages
load_requirements()
as.vector(apply(sapply(sessionInfo()$otherPkgs, \(x) x[c("Package", "Version")]), 2, \(x) paste(x, collapse = "_"))) # currently loaded packages incl. version


# load packages specified as main packages AND packages specified within the list '#statistics' in the 'requirements.txt.' within the current working directory
detach_none_base() # detach all none base packages
load_requirements(list = "#statistics")
as.vector(apply(sapply(sessionInfo()$otherPkgs, \(x) x[c("Package", "Version")]), 2, \(x) paste(x, collapse = "_"))) # currently loaded packages incl. version
}

}
\seealso{
\code{\link[utils]{.libPaths()}}
}
\author{
Simon Ress
}
\keyword{...}
