% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_libPaths.R
\name{modify_libPaths}
\alias{modify_libPaths}
\title{Modify the Search Paths for Packages}
\usage{
modify_libPaths(add = NULL, reset = NULL)
}
\arguments{
\item{add}{(chr vector | NULL): Directories to be added to the search paths. NULL: add nothing}

\item{reset}{(chr vector | NULL): Directories to which the search paths are reset. NULL: don't reset}
}
\description{
Modify the "Search Paths for Packages" (.lib.loc in .libPaths-env.),
by (1) adding new directories to the search paths or (2) resetting the search paths
to a vector of directories (and deleting all previous directories).
}
\section{Dependencies}{
 NONE
}

\section{Side effects}{
 Modify the "Search Paths for Packages" (.lib.loc in .libPaths-env.) as specified.
}

\section{Return}{
 NONE
}

\examples{
\dontrun{
#Adding the folder of dplyr version 1.1.2 to the search paths
modify_libPaths(add="C:/Users/<User>/Documents/NewProject/lib/dplyr_1.1.2")

#Resetting the search paths to a vector of directories
modify_libPaths(reset=c("C:/Users/<User>/Documents/NewProject/lib1/", "C:/Users/<User>/Documents/NewProject/lib2"))
}
}
\seealso{
`browseVignettes("pama")`, `help(package = "pama")`
}
\keyword{default}
\keyword{folder}
\keyword{library}
