% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pama-package.R
\docType{package}
\name{pama-package}
\alias{pama}
\alias{pama-package}
\title{pama: Install/Load Specific Package Versions and Use Requirement.txt-files}
\description{
Find the overview of all pama-functions here: `help(package = "pama")`

To learn more about pama, start with the vignettes: `browseVignettes("pama")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SimonRess/PaMaTo}
  \item Report bugs at \url{https://github.com/SimonRess/PaMaTo/issues}
}

}
\author{
\strong{Maintainer}: Simon Ress \email{simon.ress89@gmail.com} (\href{https://orcid.org/0000-0001-7325-7462}{ORCID})

}
\keyword{internal}
