% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMA.F.R
\name{PAMA.F}
\alias{PAMA.F}
\title{This function implements Maximum Likelihood estimation of PAMA model.}
\usage{
PAMA.F(datfile, nRe, threshold, iter = 1000, init = "EMM")
}
\arguments{
\item{datfile}{A matrix or dataframe. This is the data where our algorithm will work on. Each row denotes a ranker's ranking. The data should be in entity-based format.}

\item{nRe}{A number. Number of relevant entities.}

\item{threshold}{A number (positive). The stopping threshold in determining convergence of MLE. If the two consecutive iterations of log-likelihood is smaller than the threshold, then the convergence is satisfied.}

\item{iter}{A number. Numner of iterations of MCMC.}

\item{init}{A string. This indicates which method is used to initiate the starting point of the aggregated ranking list. "mean" uses the sample mean. "EMM" uses the method from R package 'ExtMallows'.}
}
\value{
List. It contains MLE of all the parameters and log-likelihood. We use an iterative procedure to find the MLEs, so there are several values for each parameter until convergence.
\enumerate{
\item I.mat:  samples of I
\item phi.mat:  samples of phi
\item smlgamma.mat:  samples of gamma
\item l.mat:  samples of log-likelihood
}
}
\description{
This function implements Maximum Likelihood estimation of PAMA model.
}
\examples{
a=NBANFL()
PAMA.F(a$NBA,nRe=10,threshold=0.1,iter=10)
}
\references{
Wanchuang Zhu, Yingkai Jiang, Jun S. Liu, Ke Deng (2021) Partition-Mallows Model and Its Inference for Rank Aggregation. Journal of the American Statistical Association
}
\author{
Wanchuang Zhu, Yingkai Jiang, Jun S. Liu, Ke Deng
}
