% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMAlike.R
\name{PAMAlike}
\alias{PAMAlike}
\title{this function returns the log-likelihood of an observed ranking list given the parameters.}
\usage{
PAMAlike(bsrkr, I, phi, smlgamma)
}
\arguments{
\item{bsrkr}{It is a observed ranking list.}

\item{I}{It is the true classification of entities. 0 denotes the corresponding entity is a background entity. The positive integer denotes the relative rankings of a relevant entity.}

\item{phi}{It is a positive number. It is the common disperse parameter in Mallows model for all the rankers}

\item{smlgamma}{A positive number. It is the quality parameter of the ranker. It is used to distinguish relative and background entities}
}
\value{
The lon-likelihood of barkr given I, phi and smlgamma
}
\description{
this function returns the log-likelihood of an observed ranking list given the parameters.
}
\examples{
dat=t(PerMallows::rmm(10,1:20,0.5))
I=c(1:10,rep(0,10))
like=PAMAlike(bsrkr=dat[,1],I=I,phi=0.2,smlgamma=1)
}
