% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_version.R,
%   R/update_packages_search_path.r
\name{library_version}
\alias{library_version}
\alias{update_packages_search_path}
\title{...}
\usage{
library_version(package, version, lib.search.path = .libPaths()[1])

update_packages_search_path(path = NULL)
}
\arguments{
\item{package}{(chr vector): ...}

\item{version}{(chr vector): ...}

\item{lib.search.path}{(chr vector): ...}

\item{path}{- (chr vector): Location of R library tree which should be added to the seach path}
}
\description{
...

Updating Search Paths for Packages
}
\details{
test

test
}
\section{Dependencies}{
 - update_packages_search_path()

 None
}

\section{Side effects}{
 ...

 Updating Search Paths for Packages, see .libPaths()
}

\section{Return}{
 ...

 None
}

\examples{
\dontrun{
# Load v3.4.0 of ggplot2 from folder "ggplot2_3.4.0" in search paths
library_version("ggplot2", "3.4.0")
}

\dontrun{
#Adding all version-specific packages folders within current search paths to the search paths
.libPaths()
update_packages_search_path()
.libPaths()

#Adding a specific search path
.libPaths()
update.packages.search.path("C:/Users/Simon/AppData/Local/R/win-library/4.2/ggplot2_3.1.0")
.libPaths()
}

}
\seealso{
\code{\link[utils]{.libPaths()}}

\code{\link[utils]{.libPaths()}}
}
\author{
Simon Ress
}
\keyword{Packages}
\keyword{Paths}
\keyword{Search}
\keyword{Updates}
\keyword{for}
