% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_library_version.R
\name{apply_library_version}
\alias{apply_library_version}
\title{Adjust package-loading in R scripts to the "pama-style"}
\usage{
apply_library_version(
  req.file.path = getwd(),
  req.file.name = "requirements.txt",
  lists = TRUE,
  script.paths = NULL,
  patterns = c("library\\\\(([^,)]+)\\\\)", "require\\\\(([^,)]+)\\\\)")
)
}
\arguments{
\item{req.file.path}{(chr value): Path in which to create the requirements file}

\item{req.file.name}{(chr value): Name of the requirements-file}

\item{lists}{(bool): Insert a list structure into the file? TRUE=yes, FALSE=no}

\item{script.paths}{(chr vector): Provide a path / paths to the R scripts that should be adjusted}

\item{patterns}{(chr vector): \href{https://cran.r-project.org/web/packages/stringr/vignettes/regular-expressions.html}{RegEx} to detect package-loadings in the r-scripts}
}
\description{
Adjust package-loading in R scripts to the "pama-style"
}
\details{
Adjusting the package-loading in R scripts to the "pama-style". It replaces all functions defined in `<patterns>` (default: `library()` and `require()`) by the pama-function
`library_version(<package_name>, <package_version>)`. The `<package_name>` attribute is extracted from the scripts. The `<package_version>` is read from the specified requirements-file
}
\section{Dependencies}{

- get_requirements()
}

\section{Side effects}{
 Adjusting the package-loading in R scripts to the "pama-style"
}

\section{Return}{
 None
}

\examples{
\dontrun{

#Adjust package-loading within R scripts in .\Scripts to pama-style, using versions from the "requirements.txt"-file
apply_library_version(req.file.path = r"(C:\Users\<User>\Desktop\titanic-r-master)",
                      req.file.name = "requirements.txt",
                      script.paths = r"(C:\Users\<User>\Desktop\titanic-r-master\Scripts)")
}

}
\author{
Simon Ress
}
\keyword{library_version}
\keyword{pama-style,}
