% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations2.R
\name{citations2}
\alias{citations2}
\title{Creation of a 'citation table' for packages}
\usage{
citations2(
  use.req.file = TRUE,
  req.file.name = "requirements.txt",
  req.file.path = getwd(),
  packages = c("ggplot2", "dplyr"),
  lib.loc = NULL,
  output.path = getwd()
)
}
\arguments{
\item{use.req.file}{(bool): Whether (TRUE) to use a requirements-file as source of packages to be cited or (FALSE) to provide package name in 'packages'-attribute}

\item{req.file.name}{(chr value): Name of the requirements-file}

\item{req.file.path}{(chr value): Path of the requirements-file}

\item{packages}{(chr vector): If requirements-file is not used, a vector with the names of the packages to be cited}

\item{lib.loc}{(bool | chr vector): path names of R libraries, or the directory containing the source for package, or NULL. The default value of NULL corresponds to all libraries currently known. If the default is used, the loaded packages are searched before the libraries.}

\item{output.path}{(chr value): Where to save the csv-file containing the citation table}
}
\description{
Creation of a 'citation table' for packages
}
\details{
Creating a citation table the given packages including the columns 'Package', 'Version' and 'Zitat'.
}
\section{Dependencies}{
 get_requirements()
}

\section{Side effects}{
 Saving the citation table for packages as .csv-file within the output.path. File name: paste0("/lib_citations_",Sys.Date(),".csv")
}

\section{Return}{
 Data frame including the citation table for packages
}

\examples{
\dontrun{
# Citing the packages from the "requirements.txt" within the current working directory
citations2()

# Citing the packages from the "req.txt" within "C:/GIT/Packages/PaMaTo" and saving ourput within the current working directory
citations2(req.file.path = "C:/GIT/Packages/PaMaTo",
           req.file.name = "req.txt")
}

}
\seealso{
\code{\link[pama]{get_requirements}}
}
\author{
Simon Ress
}
\keyword{citation}
\keyword{packages}
