% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reqirements_file.R
\name{create_reqirements_file}
\alias{create_reqirements_file}
\title{Create an draft requirements-file}
\usage{
create_reqirements_file(
  req.file.path = getwd(),
  req.file.name = "draft-requirements.txt",
  lists = TRUE
)
}
\arguments{
\item{req.file.path}{(chr vector): Path in which to create the requirements file}

\item{req.file.name}{(chr vector): Name of the requirements-file}

\item{lists}{(bool): Insert a list structure into the file? TRUE=yes, FALSE=no}
}
\description{
Create an draft requirements-file
}
\details{
Create an draft requirements-file which can be filled with needed packages+versions & package-lists
}
\section{Dependencies}{
 None
}

\section{Side effects}{
 Creating an draft requirements-file in the given folder
}

\section{Return}{
 None
}

\examples{
\dontrun{
#Creating a req-file in the working directory called "draft-requirements.txt", with list structure
create_reqirements_file()

#Creating a req-file in "C:/Project" called "requirements.txt", without list structure
create_reqirements_file(req.file.path="C:/Project", req.file.name="requirements.txt", lists=FALSE)
}
}
\keyword{create}
\keyword{requirements-file}
