% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_none_base.R
\name{detach_none_base}
\alias{detach_none_base}
\title{Detach all none base packages}
\usage{
detach_none_base(None)
}
\arguments{
\item{None}{Always all non base packages (except of package:PaMa itself) will be detached}
}
\description{
Detach all (none base) packages -> start with empty package-list
}
\details{
Detaches all packages listed in utils::sessionInfo()$otherPkgs except of package:pama
}
\section{Dependencies}{
 NONE
}

\section{Side effects}{
 Detaches all (none base) packages
}

\section{Return}{
 vector of attached none base packages -> is empty ("") if it worked
}

\examples{
\dontrun{
# Load requirements specifies in "requirements.txt" within the working directory
detach_none_base()
}

}
\seealso{
\code{\link[base]{.libPaths}}
}
\author{
Simon Ress (simon-ress.de)
}
\keyword{detach}
\keyword{namespace}
\keyword{non-base}
\keyword{packages}
