% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Get the dependencies of an package on CRAN}
\usage{
get_dependencies(
  package,
  version,
  cran.mirror = "https://cloud.r-project.org/",
  archiv.path = "src/contrib/Archive/",
  main.path = "src/contrib/",
  search.for.cran.name = TRUE,
  check.in.lib = FALSE
)
}
\arguments{
\item{package}{(chr value): Name of the package to install (e.g. "ggplot2")}

\item{version}{(chr value): Version of the package to install (e.g. "3.4.0")}

\item{cran.mirror}{(chr value): Main url of the cran mirror to use (e.g. "https://cloud.r-project.org/")}

\item{archiv.path}{(chr value): URL-path to the archive of the cran mirror to use (e.g. "src/contrib/Archive/")}

\item{main.path}{(chr value): URL-path to the pages main page of the cran mirror to use (e.g. "src/contrib/"")}

\item{search.for.cran.name}{(bool): Should CRAN be searched for other name-structures of the required version?}

\item{check.in.lib}{(bool): FALSE: Download the package from <cran> and extract dependencies from the "DESCRIPTION" file within the downloaded folder. TRUE: Use the "DESCRIPTION" file within the package discoverable in .libPaths()
e.g. 0.1.10 to 0.1-1 see \code{\link[pama]{find_package_version_on_cran}}
TRUE=yes, FALSE=no}
}
\description{
Extracts all dependencies (package names and versions) of the giving version of the package
}
\note{
See all available CRAN Packages by Name here: https://cran.r-project.org/web/packages/available_packages_by_name.html
}
\section{Dependencies}{

- find_package_version_on_cran()
}

\section{Side effects}{
 None
}

\section{Return}{
 List of lists with structure c(<package_name>, <version>)
}

\examples{
\dontrun{

#package-VERSION does not exist -> returns "version-error"
get_dependencies("ggmap","2.6.0" )

# package-VERSION does not exist -> returns "version-error"
get_dependencies("ggplot2","26.6.6" )

# package-NAME does not exist -> returns "package-name-error"
get_dependencies("ggmapAAA","3.0.0" )

# package-NAME does not exist -> returns "package-name-error"
get_dependencies("ggplot222","1.0.0" )

# package version in archive (-> https://cran.r-project.org/src/contrib/Archive/ggplot2/)
get_dependencies("ggplot2","3.1.1" )

# current version, date: 2023.03.08 (on main page -> https://cran.r-project.org/web/packages/ggplot2/index.html)
get_dependencies("ggplot2","3.4.1" )

# package version in archive (->https://cran.r-project.org/src/contrib/Archive/ggmap/)
get_dependencies("ggmap","3.0.0" )

# current version, date: 2023.03.08 (on main page -> https://cloud.r-project.org/web/packages/ggmap/)
get_dependencies("ggmap","3.0.1" )
}

}
\author{
Simon Ress
}
