% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_requirements.R
\name{get_requirements}
\alias{get_requirements}
\title{...}
\usage{
get_requirements(file.name = "requirements.txt", path = getwd())
}
\arguments{
\item{file.name}{(chr vector): ...}

\item{path}{(chr vector): ...}
}
\description{
...
}
\details{
test
}
\section{Dependencies}{
 NONE
}

\section{Side effects}{
 ...
}

\section{Return}{
 ...
}

\examples{
\dontrun{
# Load requirements specifies in "requirements.txt" within the working directory
get_requirements()
}

}
\seealso{
\code{\link[utils]{.libPaths()}}
}
\author{
Simon Ress
}
\keyword{...}
