% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package_version.R
\name{install_package_version}
\alias{install_package_version}
\title{....}
\usage{
install_package_version(package, version, lib.install.path = .libPaths()[1])
}
\arguments{
\item{package}{(chr vector): Name of the package to install (e.g. "ggplot2")}

\item{version}{(chr vector): Version of the package to install (e.g. "3.4.0")}

\item{lib.install.path}{(chr vector): Folder in which to install the packages}
}
\description{
!!! Recursive function !!!
}
\note{
See all available CRAN Packages by Name here: https://cran.r-project.org/web/packages/available_packages_by_name.html
}
\section{Dependencies}{

- get_installed_packages() [<- update_packages_search_path()]

- get_dependencies()

- update_packages_search_path()
}

\section{Side effects}{
 Installation of the package & adding package location to the search paths
}

\section{Return}{
 TRUE (successful installation) or FALSE (UNsuccessful installation)
}

\examples{
\dontrun{
sessionInfo()
install_package_version("ggplot2", "3.4.0")
sessionInfo()
library_version("ggplot2", "3.4.0")
sessionInfo()
detach(paste0("package:","ggplot2"), character.only = TRUE) # character.only = TRUE <- needed when paste0() or object used
}

}
\author{
Simon Ress
}
\keyword{installing}
\keyword{package-versions}
