% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package_version.R
\name{install_package_version}
\alias{install_package_version}
\title{Installing a specific version of a package (keep other versions!)}
\usage{
install_package_version(
  package,
  version,
  lib.install.path = .libPaths()[1],
  cran.mirror = "https://cloud.r-project.org/",
  archiv.path = "src/contrib/Archive/",
  main.path = "src/contrib/",
  auto.update.version.in.files = TRUE
)
}
\arguments{
\item{package}{(chr vector): Name of the package to install (e.g. "ggplot2")}

\item{version}{(chr vector): Version of the package to install (e.g. "3.4.0")}

\item{lib.install.path}{(chr vector): Folder in which to install the packages}

\item{cran.mirror}{(chr vector): Main url of the cran mirror to use (e.g. "https://cloud.r-project.org/")}

\item{archiv.path}{(chr vector): URL-path to the archive of the cran mirror to use (e.g. "src/contrib/Archive/")}

\item{main.path}{(chr vector): URL-path to the pages main page of the cran mirror to use (e.g. "src/contrib/"")}

\item{auto.update.version.in.files}{(bool):
If TRUE the version of a package in the installed files will be changed to the required version. This only happens if
it's the same version but the structure of the version name differs e.g. 0.1.10 to 0.1-1 see \code{\link[PaMa]{find_package_version_on_cran}}
When FALSE nothing happens}
}
\description{
!!! Recursive function !!!
}
\note{
See all available CRAN Packages by Name here: https://cran.r-project.org/web/packages/available_packages_by_name.html
}
\section{Dependencies}{

- find_package_version_on_cran()
- get_installed_packages() [<- update_packages_search_path()]

- get_dependencies()

- update_packages_search_path()
}

\section{Side effects}{
 Installation of the package & adding package location to the search paths
}

\section{Return}{
 TRUE (successful installation) or FALSE (UNsuccessful installation)
}

\examples{
\dontrun{
sessionInfo()
install_package_version("ggplot2", "3.4.0")
sessionInfo()
library_version("ggplot2", "3.4.0")
sessionInfo()
detach(paste0("package:","ggplot2"), character.only = TRUE) # character.only = TRUE <- needed when paste0() or object used
}

}
\author{
Simon Ress
}
\keyword{installing}
\keyword{package-versions}
