% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package_version.R
\name{install_package_version}
\alias{install_package_version}
\title{....}
\usage{
install_package_version(package, version, lib.install.path = .libPaths()[1])
}
\arguments{
\item{package}{sad}

\item{version}{sd}

\item{lib.install.path}{vv}
}
\description{
-> Recursive function <-
}
\note{
See all available CRAN Packages by Name here: https://cran.r-project.org/web/packages/available_packages_by_name.html
}
\section{Dependencies}{

- get_installed_packages() [<- update_packages_search_path()]

- get_dependencies()

- update_packages_search_path()
}

\section{Side effects}{

}

\section{Return}{

}

\examples{
\dontrun{
sessionInfo()
install_package_version("ggplot2", "3.4.0")
sessionInfo()
library_version("ggplot2", "3.4.0")
sessionInfo()
detach(paste0("package:","ggplot2"), character.only = TRUE) # character.only = TRUE <- needed when paste0() or object used
}

}
\author{
Simon Ress
}
