% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_requirements.R
\name{install_requirements}
\alias{install_requirements}
\title{Install packages by versions specified in the requirements-file}
\usage{
install_requirements(
  req.file.path = getwd(),
  req.file.name = "requirements.txt",
  lists = "all",
  library.folder.path = getwd(),
  library.folder.name = "lib"
)
}
\arguments{
\item{req.file.path}{(chr vector): Name of the requirements-file}

\item{req.file.name}{(chr vector): Folder of the requirements-file}

\item{lists}{(chr vector): Names of lists (without '#' !) specified in the requirements-file to use}

\item{library.folder.path}{(chr vector): Folder in which the subfolder "lib" should be created}

\item{library.folder.name}{(chr vector): Name of the lib-folder}
}
\description{
Install packages by versions specified in the requirements-file
}
\details{
Wrapper around get_requirements() and install_package_version(). Installs all packages from the requirements-file (+ dependencies) in a lib-folder.
}
\section{Dependencies}{

- get_requirements()

- get_installed_packages()

- install_package_version()

- detach_none_base()
}

\section{Side effects}{
 Installation of the package & adding package location to the search paths
}

\section{Return}{
 None
}

\examples{
\dontrun{
# use 'requirements.txt.' within the current working directory
install_requirements()

# use 'req.txt' within the lib folder of the current working directory
install_requirements(req.file.path=paste(getwd(),"lib",sep="/"),
                     req.file.name="requirements.txt",
                     lists="all",
                     library.folder.path=getwd(),
                     library.folder.name="lib")
}

}
\seealso{
\code{\link[base]{.libPaths}}
}
\author{
Simon Ress
}
\keyword{installing}
\keyword{package-versions}
\keyword{requirements-file}
