% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBANFL.R
\name{NBANFL}
\alias{NBANFL}
\title{Dataset of NBA and NFL.}
\usage{
NBANFL()
}
\value{
List. It contains Bayesian posterior samples of all the parameters and log-likelihood.
\enumerate{
  \item NBA: full ranking lists of NBA power ranking.
  \item NBAPL: The partial lists of NBA power ranking
  \item NFLdata: all the ranking lists of NFL data
  \item NFLcov: covariates of NFL players
}
}
\description{
Dataset of NBA and NFL.
}
\examples{
NBANFL()
}
