% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages_search_path.r
\name{update_packages_search_path}
\alias{update_packages_search_path}
\title{Updating Search Paths for Packages}
\usage{
update_packages_search_path(path = NULL, install = FALSE)
}
\arguments{
\item{path}{- (chr vector): Location of R library tree which should be added to the seach path}
}
\description{
Updating Search Paths for Packages
}
\details{
test
}
\section{Dependencies}{
 None
}

\section{Side effects}{
 Updating Search Paths for Packages, see .libPaths()
}

\section{Return}{
 None
}

\examples{
\dontrun{
#Adding all version-specific packages folders within current search paths to the search paths
.libPaths()
update_packages_search_path()
.libPaths()

#Adding a specific search path
.libPaths()
update.packages.search.path("C:/Users/Simon/AppData/Local/R/win-library/4.2/ggplot2_3.1.0")
.libPaths()
}

}
\seealso{
\code{\link[base]{.libPaths}}
}
\author{
Simon Ress
}
\keyword{Packages}
\keyword{Paths}
\keyword{Search}
\keyword{Updates}
\keyword{for}
