% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages_search_path.r
\name{update_packages_search_path}
\alias{update_packages_search_path}
\title{Updating Search Paths for Packages}
\usage{
update_packages_search_path(
  path = NULL,
  install = FALSE,
  install.path = NULL,
  print.infos = FALSE
)
}
\arguments{
\item{path}{(chr vector): Location of R library tree which should be added to the search path}

\item{install}{Set to TRUE if update_packages_search_path() is used within the installation of a package. Then it keep only the paths to the "newest" package version}

\item{install.path}{If update_packages_search_path() is used within the installation of a package, add here the path where the package was installed. It will be added to the search paths}

\item{print.infos}{(bool): Regulates output in the users console. TRUE: All internal information will be printed (useful for debugging). FALSE (by default): No or much less information will be printed.}
}
\description{
Updating Search Paths for Packages
}
\details{
test
}
\section{Dependencies}{
 None
}

\section{Side effects}{
 Updating Search Paths for Packages, see .libPaths()
}

\section{Return}{
 None
}

\examples{
\dontrun{
#Adding all version-specific packages folders within current search paths to the search paths
.libPaths()
update_packages_search_path()
.libPaths()

#Adding a specific search path
.libPaths()
update.packages.search.path("C:/Users/Simon/AppData/Local/R/win-library/4.2/ggplot2_3.1.0")
.libPaths()
}

}
\seealso{
\code{\link[base]{.libPaths}}
}
\author{
Simon Ress
}
\keyword{Packages}
\keyword{Paths}
\keyword{Search}
\keyword{Updates}
\keyword{for}
